Move matmul demo to seperate directory.

+++ sniper/interchiplet_app/interchiplet_app.cpp
@@ -19,7 +19,7 @@ namespace nsInterchiplet
 		syscall(SYSCALL_SEND_TO_GPU,dstX,dstY,srcX,srcY,a[i],size);
 	    return 1;
     }
-    syscall_return_t readGpuMessage(int64_t dstX,int64_t dstY,int64_t srcX,int64_t srcY,int *a, int size)
+    syscall_return_t readGpuMessage(int64_t dstX,int64_t dstY,int64_t srcX,int64_t srcY,int64_t *a, int64_t size)
     {
         return syscall(SYSCALL_READ_FROM_GPU,dstX,dstY,srcX,srcY,a,size);
     }
